ENG-409: cost displayed sometimes not correct
Summary:
Cost displayed depends on the first entry matching the criteria (instanceType, regionCode & tenancy). But there were multiple sub-docs matching the criteria (one for networking, where the OS type is not set, and the other for the instance cost, where the OS type is set to Linux). We are only interested in the latter.

Test Plan: Confirmed the cost in UI

Reviewers: karthik

Reviewed By: karthik

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D518