[ENG-381] Take and release locks for docdb write transaction
Summary:
We keep the locks in write_transaction_state which is the object that persists throughout raft replication and releases them on commit.

Next step: release locks early when applied the write batch in rocksdb on leader (don't need to wait for raft replication).

Again, this is a relative differential from D442 (please review!), because that has not landed yet, and I want to split subtasks.

Test Plan:
Integration tests

By tweaking I verified that deadlocks occur if no locks are released, So they are doing something. We may want to design some tests to check lock protection more thoroughly.

Reviewers: mikhail, karthik, kannan

Reviewed By: kannan

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D472