[ENG-412] Fix math error in TestSpreadReplicasEvenly in create-table-itest
Summary: The standard deviation calculation is wrong, it should divide by the number of servers, not the mean. The test passes because the number to be tested is small for hardcoded kNumServers. The wrong formula scales up linearly with kNumServers which is not what standard deviation should do. I changed kNumServers to larger values and that immediately fails the test.

Test Plan: I ran the test a bunch of times with different kNumServers and kNumTablets. The numbers are doing what I expect them to.

Reviewers: bogdan, mikhail, karthik

Reviewed By: karthik

Subscribers: kannan, karthik, jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D538