[ENG-414] CreateTable for Redis
Summary:
This diff allows Table Creation with Redis Type. Making the logic correct to use rocksdb if it's not Kudu Table, instead of checking on Key Value table.

Redis tables are not functional yet. With the read write wiring, more switch cases will be present in tablet-service, tablet, that are incorrect now.

KEY_VALUE_TABLE_TYPE will be renamed to SQL_TABLE_TYPE in a subsequent diff.

Test Plan: Added a test in create_table-itest

Reviewers: mikhail, amitanand, kannan, karthik

Reviewed By: kannan

Subscribers: kannan, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D548