[ENG-309] Switching logic in Handle()
Summary:
The details of implementing each command (Set/Get etc) that involves talking to DocDB is out of scope.
    * This sub-task focuses only on the switching logic in the server.
    * Details of each command implementation (including the plumbing with yb-client/docdb translation) will be a separate sub-task.

Test Plan:
Unit tests.
In ENG-424, more unit tests will be added to test the command(s) end to end. It should test the code/logic in this diff even more.

Reviewers: bogdan, mikhail, kannan, karthik

Reviewed By: karthik

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D574