ENG-365: Fixing master-test due to probable timing issue
Summary:
The root cause was probably a timing issue. I was not able to repro the failure in 1k
runs, but I traced the issue to the fact that the error status is defaulted to OK, so if the async
CreateTable has not managed to fail by the time we check for failure, the very first run, we could
hit this check failure.

Test Plan: './bin/repeat_unit_test.sh debug master-test MasterTest.TestInvalidPlacementInfo'

Reviewers: bharat, karthik, mikhail

Reviewed By: mikhail

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D614