ENG-463: Refactoring placement protos and code to prepare for async replica support
Summary:
Created one more level of indirection for placement data. Now we have a ReplicationInfoPB
which will contain PlacementInfoPBs for live_replicas (the ones which replica data constnatly), as
well as async_replicas (soon to be OBSERVERs) and potentially cache_replicas, if and when we add
those.

Test Plan: running tests, created cluster, ran load test against it and it worked.

Reviewers: hector, karthik, bharat

Reviewed By: bharat

Subscribers: eng, robert

Differential Revision: https://phabricator.dev.yugabyte.com/D625