Make Default Instance Type m3.medium again
Summary:
Redux form handles select states differently in that if a default value is provided and onChange handler is not called then the value of the select does not get serialized upon form submission.
https://github.com/erikras/redux-form/issues/621
This diff overcomes that limitation.

Test Plan: Go to Create universe flow, instance type should be defaulted to m3.medium and persisted on form submission.

Reviewers: ram, karthik, bharat

Reviewed By: bharat

Subscribers: bharat, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D636