ENG-529: When getting a "table already exists" error, wait for the table to finish being created
Summary:
When getting a "table already exists" error, wait for the table to finish being created.
Otherwise we can temporarily see an empty schema and assume that the creation failed,
especially during retries and low timeouts.

Another issue was that we were using the client-prepared schema for comparing partition schema
between the table specificiation we're trying to create and what already exists on the server.
However, it does not contain column ids, which we expect for the purpose of this comparison.
The solution was to use the schema received from server, which already contains column ids and
is known to be equal to the client-side schema at the point of this comparison.

Test Plan:
Run the test in ENG-529 repeatedly.

Ran it 2000 times with 8 concurrent attempts, and found no test failures,
except occasional "Address already in use" (but that's a topic for a different fix):
https://gist.github.com/mbautin/e139aea7e20e3df78a9b6cce51e6aaae

Reviewers: kannan, bharat

Reviewed By: bharat

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D749