Do not fail rpc-test when getting errno of EPROTOTYPE (41, protocol wrong type for socket) on Mac OS X
Summary:
Do not fail rpc-test when getting errno of EPROTOTYPE (41, protocol wrong type for socket) on Mac OS X.
Not exactly sure why it happens, but I guess we can live with it as Mac OS X is just a
development/test platform. There are quite a few error codes acccepted in that test already --
anything that could happen when the server suddenly closes the socket.

Example failure:
https://detective.dev.yugabyte.com/job/yugabyte-mac-phabricator-clang-debug/1461/artifact/build/debug-clang-dynamic/yb-test-logs/bin__rpc-test/TestRpc_TestServerShutsDown.log

Test Plan:
- Jenkins
- `bin/repeat_unit_test.sh debug rpc-test TestRpc.TestServerShutsDown` on Mac OS X

Reviewers: kannan

Reviewed By: kannan

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D759