Check session exists on every top level component load and destroy session and redirect to /login otherwise
Summary:
Fixed Session Check on change on top level components.
The app will now redirect to /login if the database is cleared or the user has logged out.

Test Plan:
In Tab1 - go to localhost:3000 and login
In Tab2- go to localhost:3000 , you should be logged in.

Now logout in Tab1
Go to Tab2 and switch component pages, you should be redirected to /login

Reviewers: karthik, ram

Reviewed By: ram

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D706