ENG 554: Wait for pending tasks causes failure of second master leader step down
Summary:
When a master leader steps down, the background activity state is not proactively cleaned up (this is mostly by design due to consensus service layer processing the step down and bg state tracked in master service). So there were pending (load balance add/remove) deferred tasks which were never getting issued nor removed from the pending task list in master. Causing the next time this master becomes leader to be stuck waiting for draining these pending tasks.

Added new state to tasks for tracking if they have not even been issued, so that the pending tasks can be removed from list when this master wins election and does leader related in-memory setup. Note that the task itself is not deleted as it is part of the deferred pool responsible for reissuing the rpc when its time is up.

TODO: Add a new unit test with flags to capture these sequence of events.

Test Plan:
Unit tests.

End-to-end testing in progress.

Reviewers: bogdan, karthik, hector

Reviewed By: hector

Subscribers: eng, jenkins-bot

Differential Revision: https://phabricator.dev.yugabyte.com/D818