ENG-557: ChangeMasterConfig times out
Summary:
Add a check on the leader to not step down if it started a change role - so it might have an active or committed config with a PRE_VOTER.

Added diags for initial tablet server assignment for issues like ENG-580.

Test Plan:
New unit test and existing tests.

Tested end to end using yugaware also.

Reviewers: karthik, hector

Reviewed By: hector

Subscribers: kannan, eng, jenkins-bot

Differential Revision: https://phabricator.dev.yugabyte.com/D829