ENG-564 UniverseDetail page freezes due to uncaught error in ReactBootstrapTable
Summary:
The root cause of this is that componentWillUnmount() in the RegionMap() is throwing an error in the case where LatLongBounds is not present for the case of 2 regions.
This change will actually cause the whole map to be shown in case of 3 or more regions but handles the error case.
Will fix map bounds calculation in upcoming diff,

Test Plan: Go to universeDetailPage, modals and functionalities should work as expected.

Reviewers: ram, karthik

Reviewed By: karthik

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D833