ENG-593 Expanding the number of nodes in a single az universe fails
Summary: This is because in the case of a single az universe, we were setting the region list to be an object instead of an array

Test Plan:
Create a Single AZ universe.
Wait for it to finish.
Go to universe detail page.
Test out the expand universe by clicking edit and just increasing the number of nodes and nothing else.

Reviewers: karthik, ram

Reviewed By: ram

Subscribers: eng, amitanand

Differential Revision: https://phabricator.dev.yugabyte.com/D877