ENG-581: AZ info incorrect in Universe Overview section
Summary:
When going from single region (multi AZ) to multi region and increased nodes, we were not detecting a full-copy scenario.

Fix is to pass in a `pure expand` parameter during configure nodes, so that we do full universe mode if that parameter is false.

Test Plan:
Ran create/edit+expand and ensured the zones show up correctly and nodes are actually round robined across regions and az's in them.

Reviewers: karthik

Reviewed By: karthik

Subscribers: eng

Differential Revision: http://phabricator.dev.yugabyte.com:80/D896