ENG-590: Lots of client errors in docker based expand cluster
Summary:
If a peer has a pending config and if a new leader gets elected and asks the peer to abort that update, then the consensus meta was getting stuck with the pending config. More info in the JIRA text.

Fix is to clear the pending change config in cmeta on an abort of the op.

Test Plan:
Unit tests. Create/Edit on older build, will rerun with this rebase.

Reviewers: hector

Reviewed By: hector

Subscribers: jenkins-bot, hector, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D915