Home Page map marker, on-click detail popup display
Summary:
Added popup display with all the regions, providers, and universe in the supported region group.
Currently dumping the string into a popup dialog(Check Screenshot) will let @shawn make the css look better.

{F1637}

{F1638}

Test Plan:
Go to Home page map.
Click on any of the markers in the map.
The regions, providers and universes in the markercluster need to be shown.

Reviewers: karthik, shawn, ram

Reviewed By: ram

Subscribers: shawn, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D876