[Ready For Review] ENG-563, ENG-558 Fixed Universe Table Task Progress Display
Summary:
Fixed Universe Table Task Progress Display
The reason for the task progress display getting populated with the same data is that the get request is made on componentDidMount() and for an list of such ProgressBars the subsequent GET requests are not made.
Created a multi-get tasks API endpoint, which maintains a list of all tasks by universe and taskUUID for easy querying.
{F1591}

Test Plan:
Go to location:3000.
Create more than one universe. The task progress bar for each universe should show individual task progress.

Reviewers: karthik, shawn, ram

Reviewed By: ram

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D912