[Ready To Review] Created TableList in Universe Page Component, action, reducer API controller,
Summary:
Reverting to React-Bootstrap-Table

Created Tables List in Universe Table Tab.
Added listAll() controller method to list all tables for a given universe.
The controller method currently mocks table data and return {"tableType": "...", "tableName": "..."}
Table view Implementation uses YBTable component which is not included as a part of this diff (D934)
{F1665}

Test Plan:
patch D394, patch this diff on top of that.
Use should be able to see Tables List view in the tables tab as given in screenshot.

Reviewers: shawn, ram

Reviewed By: ram

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D936