Fixed propType warning: bodyStyle needs to be an object and not a string
Summary:
Fixed propType warning : bodyStyle needs to be an object and not a string.
bodyStyle was being set as a CSS string. While this works, it throws a React propType warning .
Fixed this by adding inline styles.

Test Plan:
Go to Tasks tab in localhost:3000
There should be no console warnings on the console.

Reviewers: shawn, ram

Reviewed By: ram

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D946