ENG-629 Incorrect error when incorrect number of AZs specified
Summary:
When incorrect number of AZs are specified, the API response is an error object instead of a userIntent object.
Therefore the error that is seen is actually correct. The userIntent object is null since the API call did not return it.
The solution seems to be to disable the submit button and show visual error on the screen similar to what we do for other validation cases.

{F1699}

Test Plan:
go to localhost:3000 Create Universe.
Specify region with incorrect number of AZ's .
There should be visual error on screen and submit button should be disabled.

Reviewers: ram

Reviewed By: ram

Subscribers: kannan, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D947