ENG-680: DBImpl::WriteImpl crashes on MacOS X
Summary:
We have seen some test failures crash in DBImpl:WriteImpl (details in ENG-680).

The previous code access logs_.back() without a mutex_. Although the comments in the header file say that the mutex is not needed, there is some logic error which does seem to need the mutex. The suggested fix now lets the failing test pass across 2000 runs. Previously, it tended to fail about twice every 1000 runs.

So while no proof that the fix works, I didn't want to spend too much time on a piece of logic which YB doesn't really care about, because under YB, we don't use RocksDB WAL; instead we use the RAFT logs as the WAL.

Test Plan:
Ran the following:

~/code/yugabyte/bin/repeat_unit_test.sh debug db_dynamic_level_test DBTestDynamicLevel.MigrateToDynamicLevelMaxBytesBase --clang -n 2000

And will wait for detective results as well.

Reviewers: mikhail, timur

Reviewed By: timur

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D993