ENG-709: Fixing sys_catalog reads with bloom filter scans
Summary:
Realized work on D815 was done in parallel to switching sys_catalog to YBSQL. This caused
some unintended breaking of the read path on master restarts which some tests highlighted.

To fix this, I added a flag in the ScanSpec, but I think this might require some more discussion on
what the best fix going forward should be. On the one hand, this is a scan specific matter, as we
only have the bloom optimization for scans. On the other hand, the ScanSpec feels like it is geared
more towards filters...

Test Plan: waiting for jenkins, ran ctest locally and re-ran flaky tests.

Reviewers: timur, kannan, mikhail

Reviewed By: mikhail

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1005