Table Detail Route, Component Definition,  Tab Definition, Reducer, Action, Controller Definition,
Summary:
Created Table Detail page layout,
This contains the routing for the TableDetails.
The route is nested within the <Universe>..</Universe> component because Tables is a Child of Universe and we want current active navigation tab to be set to Universe when Table is active.
View is empty as of now.

Test Plan:
go to localhost:3000
go to any universe detail page and click on Table tab.
Click on the name of any table.
You should now be in the Table Details View.

Reviewers: ram, shawn

Reviewed By: shawn

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D951