ENG-723 Clear last YBSqlReadOp saved in SqlEnv always
Summary:
Clear last YBSqlReadOp saved in SqlEnv always
After executing a SELECT statement, SqlEnv saves the YBSqlReadOp for the caller to generate
the RowsResultResponse for return to the CQL client. It is cleared when executing a new SELECT
statement except when the query is executed against a "system" table, causing Cassandra load
tester to fail when run repeated. The YBSqlReadOp saved is now cleared always regardless of
the SELECT statement.

We should revisit the way the rows result is returned from YB SQL layer after a SELECT statement
is excuted.

[Kannan, may I ask you to review this small diff to unblock Karthik's load testing?]

Test Plan: run Cassandra load tester repeatedly.

Reviewers: neil, kannan

Reviewed By: kannan

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1025