Expect Linuxbrew to be installed in ~/.linuxbrew-yb-build and don't expect it to be added to PATH
Summary:
It is beneficial to maintain a separate tightly controlled Linuxbrew environment in a non-standard
location (e.g. `~/.linuxbrew-yb-build`) so that no unexpected packages are likely to be installed
there by accident. This requires a few changes, including setting PATH from within the build.

Additionally, this revision is enabling ccache for the third-party build.

Test Plan: build in a variety of environments

Reviewers: karthik, bogdan

Reviewed By: bogdan

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1082