Placeholder Pages for Alerts and Help
Summary:
Created Placeholder Components for Alerts and Helps tab so that there is no error on clicking on the items.
The existing AlertsList component is actually the TasksList component so renamed it to the same.

Test Plan:
Go to localhost:3000
Click on all tabs in side menu.
You should go to the corresponding page/component and no error on any tab click

Reviewers: karthik, shawn, ram

Reviewed By: ram

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1090