ENG-757 Destroyed Nodes should not be shown in Node List
Summary:
When Node state changes to Destroyed it is no longer shown in NodeList table.
Thus in a 3 node to 3 node full move, there will be 6 rows while edit is still going on and then change to 3 rows once 3 of the nodes have been destroyed.
{F2087}

Test Plan: Go to universe detail, do a full 3 node to 3 node move, at the end of edit when move has completed only three rows in the NodeDetails table showing the currently running nodes.

Reviewers: bharat, karthik, ram

Reviewed By: ram

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1092