ENG-750: Add support for 'CREATE TABLE IF NOT EXISTS'
Summary: Modified PTCreateTable::Analyze (CREATE TABLE semantics analyzer) so that if a table already exists and the statement is not a 'CREATE TABLE IF NOT EXISTS', we return an error immediately. If the statement is a  'CREATE TABLE IF NOT EXISTS' and the table already exists, then we set the private variable table_already_exists_ which will be checked by the executor and return a Status::OK().

Test Plan:
Added new tests to ybsql-create-table-test.cc to verify
1. CREATE TABLE IF NOT EXISTS can create a table
2. CREATE TABLE fails with a specific error if the table already exists
3. CREATE TABLE IF NOT EXISTS succeeds when a table already exists

Reviewers: kannan, neil, robert

Reviewed By: neil

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1107