ENG-782 : Create Redis Table failing in create universe
Summary:
The serializer somehow didn't like the update to the nested field of multi column hash schema in the create table request builder. Moved it to constructor for now.

Test Plan:
Create universe from local UI. And ensured that the table is split correctly.

Thanks @ram for helping test it out on the scheduler machine on your birthday!

Reviewers: karthik, ram

Reviewed By: ram

Subscribers: ram, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1118