Add -Werror=string-plus-int to cmake and fix string plus int concatenation.
Summary:
We had the following warning in our build:
'adding 'int64_t' (aka 'long long') to a string does not append to the string [-Werror,-Wstring-plus-int]'.
I've fixed the warning and added -Werror=string-plus-int to cmake to ensure this is an error in the future.

Test Plan: yb_build.sh

Reviewers: akashnil, mikhail, kannan

Reviewed By: kannan

Subscribers: kannan, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1158