Allow server tarball creation only on CentOS7.
Summary: Check for non-Mac OS for now, can be made just CentOS later.

Test Plan:
Ran on my mac:

16:10 $ ./yb_release
Server tarball is only expected to be built on CentOS7.

PS: We should add an option to arc/phab to ask it not to run jenkin build/tests for such script changes...

Reviewers: ram, karthik

Reviewed By: karthik

Subscribers: karthik, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1188