ENG-848 Storage type incorrect for some instances
Summary:
Some storage types are specified as Storage Volume * Volume Count.
In case of m3.medium it was correct because there is only 1 Volume.
Thus storage needs to be multiplied with volume count in all cases.
{F2356}
{F2357}

Test Plan:
Go to localhost:3000 create universe.
Try to create universe with difference instance_types .
The correct values for resources should be displayed for each instance type.

Reviewers: ram, kannan, karthik

Reviewed By: karthik

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1237