Temporary hack to purge entire rows with table level TTL.
Summary:
For customer-xyz and customer-xyz, we would like to support purging the
primary key when a TTL set. Based on offline discussions the following is the
easiest way to accomplish this.

1) For table level TTLs for customer-xyz, we will expire the entire row once the
table level TTL expires.
2) For customer-xyz, we will add a level TTL for the entire table and ensure all
inserts/updates are done with a lower TTL to ensure that all data eventually disappears.

This can be accomplished really easily by following the rule that if the
object init marker expires based on table level TTL, we should purge it.

Test Plan: Unit tests, jenking

Reviewers: karthik, akashnil, kannan, mikhail

Reviewed By: mikhail

Subscribers: pritam, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1239