Change ttl from milliseconds to seconds from user point of view.
Summary:
Originally, I introduced TTL as being specified in milliseconds by the user. This was a
misinterpretation on my part and we'd like to be compatible with cassandra's CQL where the user
enters the TTL in seconds and not milliseconds. However, internally we still store the ttl in
milliseconds.

Test Plan: jenkins

Reviewers: neil, kannan, akashnil, mikhail

Reviewed By: mikhail

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1271