ENG-926: fix for the isClusteringKey bug where it's set to false when it should be true
Summary: The "isKey" flag was being overwritten when a column was a clustering key but not a partition key. Simple fix is to only overwrite it when an overwrite is needed.

Test Plan: Created a universe through the portal, created a table through the REST api, then used yb cli to describe table and make sure the change was reflected. Will need to push the changes to S3 for the snapshot to be updated.

Reviewers: pritam, karthik

Reviewed By: karthik

Subscribers: ram, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1284