Manually add dynamic libraries that libsasl might load to the package; symlink libdb into the lib directory
Summary:
- Manually add dynamic libraries that libsasl might load to the package
- Symlink libdb into the lib directory
- Remove `/usr/lib64` from RPATH as it is not needed there. We are still packaging `libsasl2`
  with the YB package, and we're just trying to help it find anything it loads using `dlopen()`
  by creating symlinks for those dynamic libraries in directories referenced by RPATH or
  LD_LIBRARY_PATH.

Tthis still packages libsasl2 (not packaging it introduces a weird mix of libraries in system and linuxbrew-specific locations, and because linuxbrew's dynamic loader does not search for libraries in the system path by default, things break that way), but  I'm now symlinking libdb-5.3.so into the "lib" directory, which gets added to LD_LIBRARY_PATH, and that does the trick of making `dlopen()` work for that library. `dlopen()` is most likely being invoked by `libsasl2`.

Test Plan: try to run yb-master from packaged distribution

Reviewers: karthik

Reviewed By: karthik

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1300