ENG-721: Destroy tasks before we destroy CatalogManager
Summary: Currently, tasks are marked in the aborted state whenever we destroy CatalogManager, but they might live after the destruction of CatalogManager which is causing  accesses of deallocated memory. With this change we will be able to abort (which immediately forces it to call its callback method instead of waiting until the expiration of a timer) a task before we destroy CatalogManager.

Test Plan: unit tests

Reviewers: bharat, mikhail

Reviewed By: bharat, mikhail

Subscribers: bharat, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1108