[Ready for Review] Decimal implementation
Summary:
All functionality is implemented and tests are working. What's left are refactors, updating comments for readability.

I added two more encoding functions for VarInt, so we have three total (All three are needed for different Decimal encodings)

  - Comparable Encoding (Needed for exponent in the comparable Decimal encoding)
  - Two's complement encoding (Needed for both scale and mantissa of the BigDecimal encoding)
  - Digit pairs encoding (Needed for the mantissa of the comparable Decimal encoding)

Turns out that we also need binary +/- arithmetic operators on VarInt because we need to canonicalize our decimal representation. Canonicalization means adjust the scale so that mantissa is of the form 0.d0d1...dk where both d0 and dk are non-zero. The corresponding adjustment to the exponent is + or - some offset. This is why we need arithmetic. (We could convert to int64, do arithmetic, and back, but then we may have problems if the exponent doesn't fit in int64: BigDecimal actually doesn't support that, but since we can, why not, moreover there are edge cases at the boundary of int64 values where it would break if we had overflows, but BigDecimal wouldn't).

@oleg will be primary reviewer. But I also invite others to take a look as well, @hector check the API for the Decimal encoder / decoder. I am happy to go over an overview or any details in person or video chat.

Test Plan: unit tests

Reviewers: oleg, hector, kannan, mihnea, pritam.damania, robert, karthik, neil, mikhail

Reviewed By: mikhail

Subscribers: jenkins-bot, hector, oleg, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1274