[ENG-991] Speed up create table in java unit tests.
Summary:
The java unit tests would take upto 6-8 seconds for a create table statements. The reason
for this was that we had 8 tablets per tablet server and hence it would try to create 3 * 8 * 3
tablets per create table statements. I've reduced the default from 8 to 2 tablets per server which
should speed this up quite a bit. I chose 2 instead of 1 so that we don't accidentally hide bugs
which don't show up for the num_tablets = 1 case.

Test Plan: jenkins

Reviewers: robert, neil, bharat, kannan

Reviewed By: bharat, kannan

Subscribers: kannan, eng, mikhail

Differential Revision: https://phabricator.dev.yugabyte.com/D1390