ENG-685: Fixing tsan race condition
Summary:
There's a codepath in the ts_descriptor access path that has a race condition. I
previously assumed a ts_descriptor does not get updated after initially being registered. Turns out
that was a wrong assumption and we need to guard the read of placement_id with the same lock.

Test Plan: waiting on jenkins, tested on devserver though

Reviewers: mikhail

Reviewed By: mikhail

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1421