ENG-1039: Use patchelf from Linuxbrew instead of chrpath when possible to update rpath
Summary:
Use patchelf from Linuxbrew instead of chrpath when possible. The issue is that chrpath has
trouble increasing rpath size, even if there is space available (see ENG-1039), and patchelf
correctly recognizes the fact that the file used to have a longer rpath, which was later reduced,
but there is space available to extend rpath again.

Test Plan:
Run `build-support/fix_rpath.py` with this patch on a build slave where a build failed due to a
`chrpath` issue.

Reviewers: karthik

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1437