Set interpreter (dynamic linker) in fix_rpath.py
Summary:
Another fix for the build with Linuxbrew-based prebuilt third-party tarball. We need to run
`patchelf --set-interpreter` to set the dynamic linker correctly on binaries that rely on
Linuxbrew's dynamic linker. We would not need this if we built the thirdparty tarball with Linuxbrew
installed in `/var/lib/jenkins/.linuxbrew-yb-build`, something that we should start doing in the
future. Building third-party dependencies on Jenkins itself will also simplify that process.

Test Plan:
`build-support/fix_rpath.py` followed by `yb_build.sh` on a Jenkins slave where the build
failed previously.

Reviewers: karthik, bogdan

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1438