ENG-1034 Test failure in PlacementInfoUtilTest
Summary:
The test case throws a null pointer exception, because universeTaskDefinition does not have placementInfo but in the Edit case we can assume that the placementInfo will be present in payload. So the test case should not assume that placementInfo is not present in edit case.
Adding a null check in the method fixes the tests

Test Plan:
Run sbt test
Alll test cases should pass

Reviewers: karthik, ram

Reviewed By: ram

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1434