ENG-905: yql: Support for ASC/DESC on primary keys
Summary:
This review only adds the grammar rules for 'WITH CLUSTERING ORDERING BY' and its semantic analysis. It also adds a new property for columns that specifies the sorting order.
I added some tests to verify that the parsing and analysis is working as it does in Cassandra.

Test Plan: Manual

Reviewers: kannan, pritam.damania, neil

Reviewed By: neil

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1326