[ENG-1018] Add system.peers to syscatalog and set tablet locations to master.
Summary:
To implement a system.peers virtual table, we need to ensure that we create this table on
master startup and persist it in syscatalog. Apart from this, we need to ensure that this table has
only 1 tablet which is hosted on the set of master servers.

Test Plan: jenkins

Reviewers: oleg, robert, mikhail

Reviewed By: mikhail

Subscribers: bharat, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D1590