ENG-1045 Universe's Metrics tab doesn't show metrics until browser refresh
Summary:
This behaviour is being caused by the React-Dimensions plugin that we have imported into the component.
This plugin causes the universe metrics not to load and throws a constant spew of error messages.
{F2726}

{F2727}

Removed the react-dimensions plugin and added some placeholder width to achieve similar layout to what we originally had.

Similar behaviour can be achieved simply by adding a resize event listener and using a debounce utilty .
Will tackle that in a follow-up diff.

Test Plan:
Go to universe metrics page for any universe.
The corresponding universe metrics should show

Reviewers: karthik, shawn, ram

Reviewed By: ram

Subscribers: ui, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D1481