ENG-1108: Invoke SetFlag at correct server
Summary:
There is a vector of tablet servers obtained from the values of the UUID to tablet server map (essentially random order), and the RPC method SetFlag should be invoked for the first server from this list.
But it was being invoked for the first server stored in ExternalMiniCluster's vector instead, which might not match the vector mentioned above.

Test Plan:
Run:
  ybd --cxx-test raft_consensus-itest --gtest_filter RaftConsensusITest.TestUpdateConsensusErrorNonePrepared -n 100
All tests should pass

Reviewers: robert, mikhail

Reviewed By: mikhail

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D1543