Use a global backtrace_state instead of allocating a new one each time
Summary:
Creating a new backtrace_state all the time might not be a good idea -- that consumes a lot of
memory and makes GetStackTrace() quite slow. This is probably not the intended way to use
libbacktrace. One instance with "threaded" set to 1 might work better.

Test Plan:
Run GetStackTrace() in a loop and ensure the reported memory usage of the process does
not keep increasing and it is reasonably fast.

Reviewers: karthik, bharat

Reviewed By: bharat

Subscribers: jenkins-bot, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D1536