ENG-1131: TsRecoveryITest_TestRestartWithOrphanedReplicates failures
Summary:
The issue turned out not to be specific to ASAN.

Background: The test probabilistically injects failure in a tablet server (i.e. causes it to crash) during the WAL append operation.  In the test, each WAL append operation (on tablet server #0) has a 5% chance of crashing the tablet server after a WAL append. The test waits until one such failure happens, it then restarts the tablet server, and verifies that all the inserted data is readable.

However, if the TS failure happens during the create table operation, then we currently don't seem to recover from that cleanly, and the create table operation times out. I think the scope of the real fix is broader and needs the master to be able to cleanly retry table creation on a different set of nodes. This is further complicated for the RF=3 case, when we currently are down to only 2 tablet servers.

For now, to address the intermittent issue in the test, moving the failure injection point to happen only after the workload has started (in other words, after the create table step is done).

Filed ENG-1135 to track the issue of recovering gracefully from TS failures during create table.

Test Plan: Previously, the test failure about three times during 50 runs. After fix, was able to get 120 clean runs of the test.

Reviewers: bharat, mikhail, hector

Reviewed By: hector

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D1566